refactor: verifyTokenMiddleware
