+++ index.js
@@ -8,27 +8,7 @@ const jsonServer = require('json-server');
 
 const initDB = require('./database/initDB');
 const config = require('./config');
-const { verifyToken } = require('./utils/JWT');
 const authRouter = require('./routes/authRoutes');
-
-const checkAuth = (req, res, next) => {
-  try {
-    const [, token] = req.headers.authorization.split(' ');
-    const { err, decode } = verifyToken(token);
-    if (err) {
-      throw err;
-    }
-    req.body.userId = decode.userId;
-    if (req.method === 'POST' || req.method === 'PUT') {
-      req.body.updatedAt = new Date().toISOString();
-    }
-    return next();
-  } catch (err) {
-    const status = 401;
-    const message = 'Wrong access token';
-    return res.status(status).json({ status, message });
-  }
-};
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
